% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_basis.R
\name{as_mppi_basis}
\alias{as_mppi_basis}
\title{Create an mPPI basis object}
\usage{
as_mppi_basis(V, name = "basis", orthonormalize = TRUE)
}
\arguments{
\item{V}{matrix (V x r) whose columns span the desired subspace}

\item{name}{optional label for reporting}

\item{orthonormalize}{logical; orthonormalize columns via QR (default TRUE)}
}
\value{
object of class 'mppi_basis'
}
\description{
Create an mPPI basis object
}
