% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_scale.R
\name{.mppi_scale_matrix}
\alias{.mppi_scale_matrix}
\title{Internal: apply Edin-style scaling transforms}
\usage{
.mppi_scale_matrix(M, sigma, mode)
}
\description{
Internal: apply Edin-style scaling transforms
}
\keyword{internal}
