% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_scale.R
\name{.mppi_sigma_cols}
\alias{.mppi_sigma_cols}
\title{Internal: compute scale factors for component time courses}
\usage{
.mppi_sigma_cols(X)
}
\description{
Internal: compute scale factors for component time courses
}
\keyword{internal}
