% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipppi.R
\name{ipppi_iphs}
\alias{ipppi_iphs}
\title{Instantaneous phase synchrony aggregator}
\usage{
ipppi_iphs(U, pk, fs, band, tau_half = 10)
}
\description{
Instantaneous phase synchrony aggregator
}
\keyword{internal}
