% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_band.R
\name{mppi_bandpass}
\alias{mppi_bandpass}
\title{Simple bandpass using stats::filter or signal::butter if available}
\usage{
mppi_bandpass(x, low = NULL, high = 0.1, order = 2L)
}
\arguments{
\item{x}{T x V matrix}

\item{low}{highpass cutoff in (0, 0.5) cycles/TR; NULL for none}

\item{high}{lowpass cutoff in (0, 0.5); NULL for none}

\item{order}{filter order for Butterworth if 'signal' present}
}
\description{
Simple bandpass using stats::filter or signal::butter if available
}
