% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_beta.R
\name{mppi_beta_permute}
\alias{mppi_beta_permute}
\title{Permutation for β-mPPI omnibus per regressor}
\usage{
mppi_beta_permute(
  B,
  Pbeta,
  Cbeta = NULL,
  run_trial = NULL,
  blksize = NULL,
  Bperm = 999L,
  zero_diag = TRUE,
  seed = NULL,
  method = c("auto", "block_flip", "permute")
)
}
\description{
Permutation for β-mPPI omnibus per regressor
}
