% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_beta.R
\name{mppi_coerce_beta}
\alias{mppi_coerce_beta}
\title{Coerce fmrilss output to a trial x V beta matrix}
\usage{
mppi_coerce_beta(obj)
}
\arguments{
\item{obj}{matrix, data.frame with (trial, roi, beta), or list of numeric vectors}
}
\description{
Coerce fmrilss output to a trial x V beta matrix
}
