% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_neural.R
\name{mppi_compare_models}
\alias{mppi_compare_models}
\title{Compare bold vs neural-domain fits (Δ-gap, AIC difference)}
\usage{
mppi_compare_models(
  fit_bold,
  fit_neural,
  resid_bold = NULL,
  resid_neural = NULL,
  pk = NULL
)
}
\description{
Compare bold vs neural-domain fits (Δ-gap, AIC difference)
}
