% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_partial.R
\name{mppi_delta_partial}
\alias{mppi_delta_partial}
\title{Partial-correlation delta approximation}
\usage{
mppi_delta_partial(Theta0, DeltaTheta)
}
\arguments{
\item{Theta0}{baseline precision; @param DeltaTheta precision delta}
}
\description{
Partial-correlation delta approximation
}
