% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_band.R
\name{mppi_fit_multi_band}
\alias{mppi_fit_multi_band}
\title{Fit ΔΣ per frequency band and optionally combine}
\usage{
mppi_fit_multi_band(
  Y,
  X,
  psych_idx,
  bands,
  zero_diag = TRUE,
  scale = c("cov", "corr")
)
}
\arguments{
\item{bands}{list of lists: each with low, high, name}
}
\description{
Fit ΔΣ per frequency band and optionally combine
}
