% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_fuse.R
\name{mppi_fuse_time_beta}
\alias{mppi_fuse_time_beta}
\title{Precision-weighted fusion of ΔΣ maps from time and trial domains}
\usage{
mppi_fuse_time_beta(D_time, D_beta, var_time, var_beta)
}
\arguments{
\item{D_time}{VxV matrix}

\item{D_beta}{VxV matrix}

\item{var_time}{scalar variance estimate for D_time (use omnibus null var)}

\item{var_beta}{scalar variance estimate for D_beta}
}
\description{
Precision-weighted fusion of ΔΣ maps from time and trial domains
}
