% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_basis.R
\name{mppi_get_M_lag}
\alias{mppi_get_M_lag}
\title{Extract a lagged interaction matrix}
\usage{
mppi_get_M_lag(fit, k, lag = 0L)
}
\arguments{
\item{fit}{An \code{mppi_fit} object.}

\item{k}{Regressor index or name.}

\item{lag}{Integer lag τ (τ = 0 returns the contemporaneous matrix).}
}
\value{
Matrix representing the interaction at lag τ in the working space.
}
\description{
Extract a lagged interaction matrix
}
