% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_hrf_adapt.R
\name{mppi_hrf_adapt}
\alias{mppi_hrf_adapt}
\title{Combine multiple ΔΣ from an HRF basis using energy-maximizing weights}
\usage{
mppi_hrf_adapt(deltas_list, pk_norms = NULL, normalize = TRUE)
}
\arguments{
\item{deltas_list}{list of VxV slope matrices for the same psychological effect but different HRF basis columns}

\item{pk_norms}{optional vector of ||pk||^2 to normalize columns}

\item{normalize}{logical; if TRUE, normalize Gram by pk_norms}
}
\value{
list with Delta (combined), weights, Gram matrix M
}
\description{
Combine multiple ΔΣ from an HRF basis using energy-maximizing weights
}
