% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_instant.R
\name{mppi_instant_fit}
\alias{mppi_instant_fit}
\title{Instantaneous mPPI engine}
\usage{
mppi_instant_fit(
  Y,
  X,
  psych_idx,
  runs = NULL,
  prewhiten = FALSE,
  basis = NULL,
  domain = c("bold", "neural", "innovations"),
  deconv = NULL,
  center_by = c("none", "run"),
  na_action = c("omit_tr", "error"),
  scale = c("normalized", "cov", "corr"),
  ar_method = "ar",
  ar_order = "auto",
  instant = list(),
  ...
)
}
\description{
Instantaneous mPPI engine
}
\keyword{internal}
