% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_lag.R
\name{mppi_lag_select}
\alias{mppi_lag_select}
\title{Select lag by maximizing omnibus statistic with permutations}
\usage{
mppi_lag_select(
  Y,
  X,
  psych_idx,
  lags = -2:2,
  blocklens = NULL,
  B = 499L,
  blksize = 10L,
  scale = c("cov", "corr"),
  targets = NULL,
  zero_diag = TRUE
)
}
\arguments{
\item{lags}{vector of candidate lags}

\item{B}{number of permutations}

\item{targets}{optional subset of psychological regressors (names or indices)}

\item{zero_diag}{logical; zero the diagonal of returned matrices}
}
\description{
Select lag by maximizing omnibus statistic with permutations
}
