% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_adv_methods.R
\name{mppi_make_null_onsets}
\alias{mppi_make_null_onsets}
\title{Generate matched null onsets away from real events}
\usage{
mppi_make_null_onsets(Tlen, onsets_idx, n = length(onsets_idx), guard = 8L)
}
\arguments{
\item{Tlen}{number of TRs}

\item{onsets_idx}{observed event indices}

\item{n}{number of null events}

\item{guard}{guard width in TRs}
}
\value{
sorted integer vector of null onset indices
}
\description{
Generate matched null onsets away from real events
}
