% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_neural.R
\name{mppi_neural_from_fit}
\alias{mppi_neural_from_fit}
\title{Convert an existing bold-domain fit to neural domain}
\usage{
mppi_neural_from_fit(
  fit,
  X,
  psych_idx,
  h,
  K = NULL,
  method = c("gcv", "fixed"),
  lambda = 0.1,
  groups = NULL
)
}
\description{
Convert an existing bold-domain fit to neural domain
}
