% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_inference.R
\name{mppi_permute}
\alias{mppi_permute}
\title{Edgewise permutations with optional studentization (experimental)
Returns p-values matrix for a given regressor index}
\usage{
mppi_permute(
  fit,
  k = 1L,
  blksize = 10L,
  B = 999L,
  wild = c("none", "rademacher", "mammen"),
  studentize = TRUE,
  method = c("block_flip", "phase", "freedman_lane"),
  seed = NULL
)
}
\description{
Edgewise permutations with optional studentization (experimental)
Returns p-values matrix for a given regressor index
}
