% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_basis.R
\name{mppi_reconstruct_delta}
\alias{mppi_reconstruct_delta}
\title{Reconstruct a submatrix of Δ from a basis fit}
\usage{
mppi_reconstruct_delta(fit, k, rows = NULL, cols = NULL)
}
\description{
Reconstruct a submatrix of Δ from a basis fit
}
