% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_utils.R
\name{mppi_select_psych}
\alias{mppi_select_psych}
\title{Select psychological regressors from a design matrix}
\usage{
mppi_select_psych(X, patterns = NULL, names_or_idx = NULL)
}
\arguments{
\item{X}{design matrix (T x q)}

\item{patterns}{character vector of regex patterns (applied to colnames(X))}

\item{names_or_idx}{specific names or integer indices}
}
\value{
integer vector of column indices
}
\description{
Select psychological regressors from a design matrix
}
