% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_adv_methods.R
\name{mppi_theta_metrics}
\alias{mppi_theta_metrics}
\title{Baseline precision metrics}
\usage{
mppi_theta_metrics(Theta0)
}
\arguments{
\item{Theta0}{precision matrix (positive definite)}
}
\value{
list with algebraic connectivity, effective rank, Qmax, and mean communicability
}
\description{
Baseline precision metrics
}
