% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_partial.R
\name{mppi_to_partial}
\alias{mppi_to_partial}
\title{First-order precision update from covariance slope}
\usage{
mppi_to_partial(Sigma0, DeltaSigma, lambda = 0.001)
}
\arguments{
\item{Sigma0}{baseline covariance (T^{-1} R'R) or correlation}

\item{DeltaSigma}{slope matrix for regressor k}

\item{lambda}{ridge parameter for invertibility}
}
\value{
matrix ΔΘ
}
\description{
First-order precision update from covariance slope
}
