% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_fit.R
\name{mppi_fit_from_fmrireg}
\alias{mppi_fit_from_fmrireg}
\title{Convenience wrapper when using fmrireg objects}
\usage{
mppi_fit_from_fmrireg(
  fmodel,
  dataset,
  psych,
  scale = c("normalized", "cov", "corr"),
  zero_diag = TRUE,
  runs = NULL,
  center_by = c("none", "run"),
  na_action = c("omit_tr", "error"),
  backend = c("blas", "accumulate", "chunked"),
  chunk_size = NULL,
  packed = FALSE,
  basis = NULL,
  project_backend = c("blas", "chunked"),
  project_chunk_cols = NULL,
  domain = c("neural", "bold", "innovations"),
  deconv = NULL,
  lags = 0L,
  lag_blocklens = NULL
)
}
\arguments{
\item{fmodel}{fmri_model (from fmrireg)}

\item{dataset}{fmrireg dataset supplying Y}

\item{psych}{regex patterns or indices}

\item{scale}{"normalized", "cov", or "corr"}
}
\description{
Convenience wrapper when using fmrireg objects
}
