% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mppi_viz.R
\name{mppi_modes}
\alias{mppi_modes}
\title{Leading eigenmodes of ΔΣ}
\usage{
mppi_modes(Delta, r = 3L)
}
\arguments{
\item{Delta}{matrix VxV}

\item{r}{number of modes}
}
\value{
list with values and vectors
}
\description{
Leading eigenmodes of ΔΣ
}
